refactor: shrink row_decoder/mod.rs below its pre-change size; dedupe the error mapper (#3095)
The file-size ratchet caught row_decoder/mod.rs growing 1076 -> 1080 (the new
timestamp_policy module declaration). Rather than waive it, split the row DISPLAY
+ static-merge helpers (row_has_non_key_cell, merge_static_cells,
build_display_row, extract_clustering_values — one shared decision each) into
display_row.rs: mod.rs 1076 -> 990. Both over-threshold core files now SHRINK
relative to main (block_emit_windowed.rs 1075 -> 972).

Also: bypass.rs's map_scan_error was a byte-identical duplicate of
row_source::map_core_error — one mapping site now, shared by both arms; and
StaticMergeSource::emitted_clustering_row documents why 'handed downstream'
(rather than 'survived the predicate') is the Cassandra-faithful test: a predicate
can only drop clustering rows when it restricts a clustering/regular column, and
in exactly that case returnStaticContentOnPartitionWithNoRows() is false.
